+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/debugger/relay/Relay.kt
@@ -1,0 +1,10 @@
+/*
+ * Copyright 2018-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
+ */
+
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.channels.SendChannel
+
+interface Relay {
+    suspend fun process(receiveFrom: ReceiveChannel<ByteArray>, sendTo: SendChannel<ByteArray>)
+}